invoiceiq/services/api/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
pydantic==2.9.2
SQLAlchemy==2.0.36
mysqlclient==2.2.4
python-multipart==0.0.9
Pillow==11.0.0
numpy==2.1.3
pdf2image==1.16.3
paddleocr==2.7.3
paddlepaddle==2.6.1
pytest==8.1.1
